refactoring features to separate BED and GFF3 packages
