+++ convert/src/test/java/org/dishevelled/bio/convert/BedRecordToFeatureTest.java
@@ -44,7 +44,7 @@ import org.bdgenomics.convert.bdgenomics.BdgenomicsModule;
 import org.bdgenomics.formats.avro.Feature;
 import org.bdgenomics.formats.avro.Strand;
 
-import org.dishevelled.bio.feature.BedRecord;
+import org.dishevelled.bio.feature.bed.BedRecord;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
