+++ convert/src/test/java/org/dishevelled/bio/convert/FeatureToBedRecordTest.java
@@ -47,8 +47,8 @@ import org.bdgenomics.convert.bdgenomics.BdgenomicsModule;
 import org.bdgenomics.formats.avro.Feature;
 import org.bdgenomics.formats.avro.Strand;
 
-import org.dishevelled.bio.feature.BedFormat;
-import org.dishevelled.bio.feature.BedRecord;
+import org.dishevelled.bio.feature.bed.BedFormat;
+import org.dishevelled.bio.feature.bed.BedRecord;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
