+++ convert/src/test/java/org/dishevelled/bio/convert/Gff3RecordToFeatureTest.java
@@ -46,7 +46,7 @@ import org.bdgenomics.formats.avro.Feature;
 import org.bdgenomics.formats.avro.OntologyTerm;
 import org.bdgenomics.formats.avro.Strand;
 
-import org.dishevelled.bio.feature.Gff3Record;
+import org.dishevelled.bio.feature.gff3.Gff3Record;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
