+++ feature/src/main/java/org/dishevelled/bio/feature/bed/BedListener.java
@@ -21,7 +21,7 @@
     > http://www.opensource.org/licenses/lgpl-license.php
 
 */
-package org.dishevelled.bio.feature;
+package org.dishevelled.bio.feature.bed;
 
 /**
  * BED listener.
