+++ feature/src/main/java/org/dishevelled/bio/feature/gff3/package-info.java
@@ -1,0 +1,30 @@
+/*
+
+    dsh-bio-feature  Sequence features.
+    Copyright (c) 2013-2020 held jointly by the individual authors.
+
+    This library is free software; you can redistribute it and/or modify it
+    under the terms of the GNU Lesser General Public License as published
+    by the Free Software Foundation; either version 3 of the License, or (at
+    your option) any later version.
+
+    This library is distributed in the hope that it will be useful, but WITHOUT
+    ANY WARRANTY; with out even the implied warranty of MERCHANTABILITY or
+    FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public
+    License for more details.
+
+    You should have received a copy of the GNU Lesser General Public License
+    along with this library;  if not, write to the Free Software Foundation,
+    Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA.
+
+    > http://www.fsf.org/licensing/licenses/lgpl.html
+    > http://www.opensource.org/licenses/lgpl-license.php
+
+*/
+
+/**
+ * Sequence features in GFF3 format. Refactored from package org.dishevelled.bio.feature.
+ *
+ * @since 2.0
+ */
+package org.dishevelled.bio.feature.gff3;
