+++ tools/src/main/java/org/dishevelled/bio/tools/CompressBed.java
@@ -32,10 +32,10 @@ import java.io.PrintWriter;
 
 import java.util.concurrent.Callable;
 
-import org.dishevelled.bio.feature.BedListener;
-import org.dishevelled.bio.feature.BedReader;
-import org.dishevelled.bio.feature.BedRecord;
-import org.dishevelled.bio.feature.BedWriter;
+import org.dishevelled.bio.feature.bed.BedListener;
+import org.dishevelled.bio.feature.bed.BedReader;
+import org.dishevelled.bio.feature.bed.BedRecord;
+import org.dishevelled.bio.feature.bed.BedWriter;
 
 import org.dishevelled.commandline.ArgumentList;
 import org.dishevelled.commandline.CommandLine;
