+++ tools/src/main/java/org/dishevelled/bio/tools/FilterGff3.java
@@ -47,10 +47,10 @@ import com.google.common.collect.Range;
 
 import org.dishevelled.bio.range.Ranges;
 
-import org.dishevelled.bio.feature.Gff3Listener;
-import org.dishevelled.bio.feature.Gff3Reader;
-import org.dishevelled.bio.feature.Gff3Record;
-import org.dishevelled.bio.feature.Gff3Writer;
+import org.dishevelled.bio.feature.gff3.Gff3Listener;
+import org.dishevelled.bio.feature.gff3.Gff3Reader;
+import org.dishevelled.bio.feature.gff3.Gff3Record;
+import org.dishevelled.bio.feature.gff3.Gff3Writer;
 
 import org.dishevelled.commandline.ArgumentList;
 import org.dishevelled.commandline.CommandLine;
