+++ tools/src/main/java/org/dishevelled/bio/tools/IntersectBed.java
@@ -49,10 +49,10 @@ import com.google.common.collect.Range;
 import com.google.common.collect.RangeSet;
 import com.google.common.collect.TreeRangeSet;
 
-import org.dishevelled.bio.feature.BedListener;
-import org.dishevelled.bio.feature.BedReader;
-import org.dishevelled.bio.feature.BedRecord;
-import org.dishevelled.bio.feature.BedWriter;
+import org.dishevelled.bio.feature.bed.BedListener;
+import org.dishevelled.bio.feature.bed.BedReader;
+import org.dishevelled.bio.feature.bed.BedRecord;
+import org.dishevelled.bio.feature.bed.BedWriter;
 
 import org.dishevelled.bio.range.tree.RangeList;
 import org.dishevelled.bio.range.tree.RangeTree;
