+++ tools/src/main/java/org/dishevelled/bio/tools/SplitBed.java
@@ -31,10 +31,10 @@ import java.io.IOException;
 
 import com.google.common.io.Files;
 
-import org.dishevelled.bio.feature.BedListener;
-import org.dishevelled.bio.feature.BedReader;
-import org.dishevelled.bio.feature.BedRecord;
-import org.dishevelled.bio.feature.BedWriter;
+import org.dishevelled.bio.feature.bed.BedListener;
+import org.dishevelled.bio.feature.bed.BedReader;
+import org.dishevelled.bio.feature.bed.BedRecord;
+import org.dishevelled.bio.feature.bed.BedWriter;
 
 import org.dishevelled.commandline.ArgumentList;
 import org.dishevelled.commandline.CommandLine;
