+++ tools/src/main/java/org/dishevelled/bio/tools/SplitGff3.java
@@ -31,10 +31,10 @@ import java.io.IOException;
 
 import com.google.common.io.Files;
 
-import org.dishevelled.bio.feature.Gff3Listener;
-import org.dishevelled.bio.feature.Gff3Reader;
-import org.dishevelled.bio.feature.Gff3Record;
-import org.dishevelled.bio.feature.Gff3Writer;
+import org.dishevelled.bio.feature.gff3.Gff3Listener;
+import org.dishevelled.bio.feature.gff3.Gff3Reader;
+import org.dishevelled.bio.feature.gff3.Gff3Record;
+import org.dishevelled.bio.feature.gff3.Gff3Writer;
 
 import org.dishevelled.commandline.ArgumentList;
 import org.dishevelled.commandline.CommandLine;
